refactor: MarkdownController -> /controllers
