+++ src/components/UI/HeadersDropdown/HeadersDropdown.tsx
@@ -4,7 +4,7 @@ import '@reach/tooltip/styles.css'
 import cn from 'classnames'
 import { FC, useRef } from 'react'
 import { useThemeContext } from '../../../context/ThemeContext'
-import MarkdownController from '../../../helpers/MarkdownController'
+import MarkdownController from '../../../controllers/MarkdownController'
 import { useClickOutside } from '../../../hooks/useClickOutside'
 import { Icons } from '../../../utils/Icons'
 import { headers } from '../../../utils/headers'
