+++ src/components/UI/MarkdownContainer/MarkdownContainer.tsx
@@ -7,7 +7,7 @@ import ActionButton from '../ActionButton/ActionButton'
 import { Icons } from '../../../utils/Icons'
 import { useRef, useState } from 'react'
 import { useMarkdownContext } from '../../../context/MarkdownContext'
-import MarkdownController from '../../../helpers/MarkdownController'
+import MarkdownController from '../../../controllers/MarkdownController'
 import ControlsContainer from '../ControlsContainer/ControlsContainer'
 import HeadersDropdown from '../HeadersDropdown/HeadersDropdown'
 import download from 'react-useanimations/lib/download'
